resultsview updates the display directly
