+++ src/index.ts
@@ -1,12 +1,11 @@
 import "../style/styles.scss";
 import { EventEmitter } from "./core/EventEmitter";
 import { IPokemon } from "./model/Pokemon";
-import "./view/AutoCompleteView";
 import { searchInAPI } from "./service/SearchService";
 import { button__next_pokemon, button__next_view, button__previous_pokemon, button__previous_view, button__search } from "./ui/Buttons";
 import { input__search } from "./ui/DomElements";
+import "./view/AutoCompleteView";
 import { ResultsView } from "./view/ResultsView";
-import { renderWithError } from "./view/updateViewFunctions";
 
 let lastSearchedPokemon: undefined | IPokemon;
 
@@ -39,5 +38,5 @@ button__previous_pokemon.addEventListener("click", () => {
 
 // Listning to events
 EventEmitter.on("search", (searchParam: string) => {
-    searchInAPI(searchParam).then(saveSearch).then(ResultsView.renderView).catch(renderWithError);
+    searchInAPI(searchParam).then(saveSearch).then(ResultsView.renderView).catch(ResultsView.renderWithError);
 })
